chore: Add dummyimage.com to allowed image domains
This commit adds 'dummyimage.com' to the list of allowed image domains in the Next.js configuration file. This will ensure that images from 'dummyimage.com' can be loaded and displayed correctly in the application.

Note: This commit message follows the established convention of starting with a verb in the imperative mood, followed by a brief description of the changes.